Foundation for implementing continuations independent of Loom.
