+++ substratevm/src/com.oracle.svm.core/src/com/oracle/svm/core/SubstrateOptions.java
@@ -106,8 +106,18 @@ public class SubstrateOptions {
         }
     };
 
-    @Option(help = "Build with Loom JDK") //
-    public static final HostedOptionKey<Boolean> UseLoom = new HostedOptionKey<>(false);
+    @Option(help = "Support continuations (without requiring a Project Loom JDK)") //
+    public static final HostedOptionKey<Boolean> SupportContinuations = new HostedOptionKey<>(false);
+
+    @Option(help = "Build with Project Loom JDK") //
+    public static final HostedOptionKey<Boolean> UseLoom = new HostedOptionKey<>(false) {
+        @Override
+        protected void onValueUpdate(EconomicMap<OptionKey<?>, Object> values, Boolean oldValue, Boolean newValue) {
+            if (newValue) {
+                SupportContinuations.update(values, true);
+            }
+        }
+    };
 
     public static final int ForceFallback = 10;
     public static final int Automatic = 5;
